Refactor package of optimize.context
